### 2021-08-06 #3579 @whedon
[PRE REVIEW]: Synaptus: A Matlab/Octave toolbox for synthetic aperture ultrasound imaging
**Submitting author:** @mh-skjelvareid (<a href="http://orcid.org/0000-0003-2103-865X">Martin H. Skjelvareid</a>)
**Repository:** <a href="https://github.com/mh-skjelvareid/synaptus" target ="_blank">https://github.com/mh-skjelvareid/synaptus</a>
**Version:** v1.1.0
