### 2021-08-06 #3579 @arfon
:wave: @mh-skjelvareid. Thanks for your submission to JOSS. Could you make some updates to your paper to make the perspective of the writing the same? Towards the end of the paper you write:

> Although the toolbox started out as a summary of my PhD work, and I have since started
> working on other research topics, I would like to keep the toolbox a “live” software project.
> I am particularly interested in adapting the algorithms in the toolbox to different imaging
> setups, e.g. different shaped arrays or imaging geometries.

To keep this consistent with the rest of the paper please change this to e.g., _Although the toolbox started out as a summary of a PhD program, efforts have since begun focused on other research topics,_ (hopefully you get the idea).
