### 2021-08-09 #3579 @mh-skjelvareid
I've looked through the list of reviewers and found a few with a suitable background (matlab, ultrasound, signal processing) - I'm listing them here without tagging them:

- lucaferranti
- benjaminbolling
- AnantHariharan1996
- astanziola
- Scivision
