### 2021-08-12 #3579 @arfon
@mh-skjelvareid – apologies for not communicating this earlier but we're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
