### 2021-08-19 #3579 @Kevin-Mattheus-Moerman
@mh-skjelvareid

Here are some recommendations:
- [ ] Change the `_readme.txt` file to a `README.md` markdown file. This way GitHub will recognize it and render it on the project front page. This helps folks clearly see what the project is about when they land on your project page.
- [ ] Change the `_license.txt` file to a plain `LICENSE` file. This way GitHub will recognize it and render a license badge at the top of your repository. This helps folks instantly see the license type for the project when they land on your page.
- [ ] Change the `_contributing.txt` file to a formatted `CONTRIBUTING.md` file and link to it from the readme.
- [ ] Furthermore, in relation to your contributor community we recommend (not required) one adds a `COC.md` (code of conduct) too. Many developers use this template: https://www.contributor-covenant.org/version/2/1/code_of_conduct/code_of_conduct.md (do not forget to amend the contact details in the template if you use it).
- [ ] If you have  a couple of example images that help show the functionality I recommend you render these in your readme file as well as in the paper. You can also consider creating a "graphical abstract" figure that summarizes functionality. This will help folks to visually grasp what the code offers.

As an example, here is a link to one of my projects: https://github.com/gibbonCode/GIBBON, note the use of the `README.md`, the way a license badge is rendered, the inclusion of a graphical overview image, the linking to contributing guidelines, and a code of conduct. You may also like the way I render "badges" in the readme for the license etc. you can copy the syntax from the readme if you want to use this too.

Quick comments on paper (we may have more later):
- [ ] In `“raw data,”` it looks like the comma should be outside of the quotation marks e.g. `“raw data”, `
- [ ] Check typo: `ultrasund` -> `ultrasound`
