### 2021-08-19 #3579 @lassoan
The topic is not exactly in my area of expertise and would require me to invest significant amount of time, which I cannot justify for a GPL-licensed software (that cannot be used freely but may require to buy commercial license for certain uses).
