### 2021-08-20 #3579 @mh-skjelvareid
@lassoan I'm quite new to publishing open source software, and I chose a license that seemed to fit, perhaps without knowing all the consequences. It was not my intention that anyone should need to buy commercial licenses to use the software (with Octave being a free alternative to Matlab). If you have suggestions for alternative licenses, I'll be happy to consider them!
