+++ experimental/msmlok2d/msmlok2d.m
@@ -1,217 +1,0 @@
-function varargout = msmlok2d(ptx,fs,mwb,cc,thick,fLow,fHigh,xsStep,xrStep,xsrOffset,varargin)
-% MSMLOK2D - omega-k migration for 2D multistatic data (full-matrix array)
-% 
-%	Usage:
-%	[im,...] = msmlok2d(ptx,fs,mwb,c,fLow,fHigh,xStep,...)
-%
-%	Input parameters:
-%   ptx			-	p(t,xr,xs): 3D matrix with ultrasonic data
-%	fs			-	Sampling frequency [Hz]
-%	mwb			-	Measurement Window Begin [s] - time delay between
-%					pulse transmission and beginning of data logging
-%	cc			-	Wave velocity [m/s]
-%	thick		-	Vector of thickness for each layer [m/s] (set last boundary equal to end of ROI)
-%	fLow		-	Lower limit of transducer frequency band [Hz]
-%	fHigh		-	Higher limit of transducer frequency band [Hz]
-%	xsStep		-	x step between source positions [m]
-%	xrStep		-	x step between receiver positions [m]
-%	xsrOffset	-	offset between first source position and first receiver position
-%
-%	Optional parameter-value input pairs:
-%	xStart		-	Lowest receiver position [m]. Default: 0. Only relevant for "xIm" output
-%	ipMethod	-	interpolation method [string] (see INTERP1 ) Default: 'linear'
-%	hh			-	Transducer impulse response, sampled at fs. Used in matched filter. Default: 1 (no filt.)
-%	tFftMult	-	multiplier for FFT size in t dir. Default: 1
-%	xFftMult	-	multiplier for FFT size in both x dir. Default: 2
-%	zFftMult	-	multiplier for z FFT size. Default: 1 
-%   ipFftMult   -   multiplier for interp. to higher resolution before interpolation from omega to kz
-%	zRes		-	Z resolution in the final image. The default resolution
-%					is given by (min(cc)/2)/fHigh, but if specified, the
-%					final image is interpolated to a resolution of zRes.
-%
-%	Output parameters (varargout):
-%	im			-	Reconstructed image
-%	xIm			-	x positions for pixels in im
-%	zIm			-	z positions for pixels in im
-%
-%	2010-01-21	Martin H. Skjelvareid
-
-%	A note on code notation:
-%	------------------------
-%	In the process of doing this transformation, many versions of the
-%	wavefield are used. A short explanation of the code notation is
-%	included here to make reading of the code easier.
-%
-%	Code variable			Represents wave field
-%	ptx						p(t,xr,xs)
-%	Pox						P(omega,xr,xs)
-%	Pokx					P(omega,kxr,kxs)
-%	Pkzkx					P(kz,kx,kxs)
-%	Pkzx					P(kz,x)
-%	pzx						p(z,x) (Image)
-
-
-%% Parse optional input parameters
-p = inputParser;
-p.addParamValue('xStart',0)							% Lowest receiver x value
-p.addParamValue('ipMethod','linear');				% Interoplation method
-p.addParamValue('xFftMult',2);						% Multiplier for FFT size in x dir.
-p.addParamValue('tFftMult',1);						% Multiplier for FFT size in x dir.
-p.addParamValue('ipFftMult',8)						% Multiplier for interpolation FFT size
-p.addParamValue('zFftMult',1)						% Multiplier for FFT size in z dir.
[…]

[…]
