### 2021-08-20 #3579 @lassoan
GPL license essentially requires all software that uses your software to be open-source. There may be exceptions depending on how exactly use the GPL code, but it is a complicated discussion that most companies want to entirely avoid. Even if you work at an academic institution, you may want your work make an impact and adopted by companies, or by other academic groups who may want to work with companies.

Apache and MIT are probably the two most popular licenses that do not restrict usage of the software in any way. Apache gives additional peace of mind to users that you don't want to make them use your patented ideas that you can later charge them for.

Unfortunately, Octave itself comes with GPL license, so even if your code has non-restrictive license, many people will still not be able to use it. Python and most Python packages comes without such restrictions.
