### 2021-08-20 627ac2c
Moved "contributing" and "license" files to markdown files, added code of conduct file.
