### 2021-08-20 540ff72
Delete _contributing.txt
