### 2021-08-20 23e2533
Added additional examples from thesis, and put them further up in the README file.
