### 2021-09-11 #3579 @Kevin-Mattheus-Moerman
@btreeby @dannyramasawmy @jarosjir @thewtex @pharshalp @scivision would you be interested in reviewing this work (a MATLAB/octave toolbox for ultrasound imaging) for the Journal of Open Source Software (JOSS)?
The smooth review process takes place on GitHub, and focuses on the software as well as a short paper.

Thanks!

Note under the current COVID19 pandemic we can be quite flexible in terms of giving your time to complete the review. If you are able to review this work in principle but currently lack the time to do so, feel free to indicate a later time when you may or estimate how long it would take to complete.

Note this is a pre-review issue in which reviewers are assigned. Once sufficient reviewers are recruited a dedicated review issue is initiated. In that review issue the reviewers are guided using check lists. Here is an example of a recently completed review issue to give you an idea of what it is like to review for JOSS: #2548
