### 2021-09-11 #3579 @Kevin-Mattheus-Moerman
> GPL license essentially requires all software that uses your software to be open-source. There may be exceptions depending on how exactly use the GPL code, but it is a complicated discussion that most companies want to entirely avoid. Even if you work at an academic institution, you may want your work make an impact and adopted by companies, or by other academic groups who may want to work with companies.
>
> Apache and MIT are probably the two most popular licenses that do not restrict usage of the software in any way. Apache gives additional peace of mind to users that you don't want to make them use your patented ideas that you can later charge them for.
>
> Unfortunately, Octave itself comes with GPL license, so even if your code has non-restrictive license, many people will still not be able to use it. Python and most Python packages comes without such restrictions.

@lassoan thanks for your comments on licenses. We respect your comments and understand them, and also that you feel you do not wish to review this work. Of course the license chosen is up to the developers of the software. At JOSS we consider works open source only if they have an [OSI approved license](https://opensource.org/licenses).

I see your points on the GPL license. These are valid concerns and certainly GPL can create compatibility issues when using/integrating the work with other software. Interestingly we have also seen reviewers strongly advocate for the GPL license and these had strong feelings against permissive licenses. Proponents of GPL might say for instance that they do not want companies to absorb and commercialize codes and then not release future developments open source. These may also be valid concerns. Either way, as long as it is an OSI approved license the work can be considered for JOSS. As a sidenote, I believe it can be challenging to enforce GPL for MATLAB works, since GPL's "midas touch", as far as I'm aware, would kick in uppon "distribution" of the software. However with MATLAB codes, if one develops codes that use this library one could always simply distribute only ones own code and ask users to download this library as a 3rd party requirement. This circumvents the distribution issue.... Anyway, this is not the place to have an extensive discussion on this. Thanks again for your comments.
