### 2021-09-11 #3579 @Kevin-Mattheus-Moerman
@mh-skjelvareid if based the above you have reconsidered your license let us know, again as long as it is an [OSI approved](https://opensource.org/licenses) open source license your work is compatible with JOSS.
