### 2021-09-13 #3579 @mh-skjelvareid
@Kevin-Mattheus-Moerman @lassoan - excuse me for not replying earlier regarding the license. And thanks, @lassoan, for your insightful comments on licensing options.

To me, the main goal of sharing this open source project is for other researchers to get a working version of a set of algorithms. Having a general, mathematical description in a paper is one thing, but an implementation reveals all the details that there's no room for in the paper. As a reasearcher, I've found that there are hundreds of papers suggesting new/improved algorithms for ultrasound imaging, but that very few authors actually publish code along with their papers. In this way, it's hard to build upon other researchers' work without spending a lot of time making your own implementation - and as a consequece, progress in the field is slow.

If a researcher decides to use one of my algorithms and maybe improve upon it, then it's important to me that the improvements are shared with the rest of the community, and this is why I've chosen the GPL license.

I can see that in some cases, GPL can be restrictive to a company. However, ultrasound imaging is very computationally intensive, and most commercial products use specially designed hardware to accelerate the process. I can't imagine that a company would use Matlab or Octave for this - a more realistic scenario is that they implement the algorithms from scratch (in e.g. C++), based on the original papers and on the _general approach_ outlined by the Matlab/Octave code. To me it seems that this would allow a company to make an implementation without being overly restricted by the GPL license. I might be wrong here - please correct me if I am.

Finally, Python seems like a very good and perhaps more "modern" alternative to Matlab, and I hope to someday make a Python implementation of the algorithms in the toolbox. However, it will require quite a lot of time, which I really don't have at the moment. I feel that publishing a Matlab version and sharing it with the community is better than waiting (potentially a few years) until a Python version is ready.
