### 2021-09-13 037fce4
Updated after comments from Kevin
Updated both README and the paper itself. Included figures in both (without captions in the paper), and used HTML to control figure sizes.
