+++ paper/paper.bib
@@ -120,6 +120,17 @@
   url    = {https://hdl.handle.net/10037/4649},
 }
 
+@article{Skjelvareid2012c,
+	author = {Skjelvareid,Martin H.  and Olofsson,Tomas  and Birkelund,Yngve },
+	title = {Three-dimensional ultrasonic imaging in multilayered media},
+	journal = {AIP Conference Proceedings},
+	volume = {1433},
+	number = {1},
+	pages = {169-172},
+	year = {2012},
+	doi = {10.1063/1.3703163}
+}
+
 @article{Skjelvareid2013,
   title={Internal pipeline inspection using virtual source synthetic aperture ultrasound imaging},
   author={Skjelvareid, Martin H and Birkelund, Yngve and Larsen, Yngvar},
@@ -187,4 +198,12 @@
   year={2017},
   organization={IEEE},
   doi={10.1109/ULTSYM.2017.8092389}
+}
+
+@misc{beyond_epica,
+  author = {Beyond EPICA team},
+  title = {Beyond EPICA - Oldest ice},
+  year = {2021},
+  journal = {Beyond EPICA website},
+  url = {https://www.beyondepica.eu/en/}
 }
