### 2021-09-13 307d291
Testing ways of displaying figures and captions
