### 2021-09-13 ddb85ea
Switched to markdown syntax for including figures.
