+++ paper/paper.md
@@ -61,38 +61,26 @@ A number of open source toolboxes related to ultrasound imaging are publicly ava
 Although the toolbox was first published as a collection of algorithms and datasets developed during a PhD program, the repository is intended to be an open and live development project. Contributions in the form of datasets from new imaging geometries (e.g. differently shaped arrays or layered media) are particularly welcome, as they provide the foundation for developing algorithms for these geometries. However, contributions in the form of code, feature requests or issue reports are all very much appreciated. Given the popularity and availability of Python, a Python implementation of the toolbox algorithms is also seen as a natural continuation of the project.
 
 # Example use
-The following example was originally presented in [@Skjelvareid2012c], and is included here to give a more visual and intuitive sense of how `Synaptus` can create a focused image from raw ultrasonic data. For the sake of brevity, only a short description is given here, but the full details of the experiment can be found in [@Skjelvareid2012c] and [@Skjelvareid2012b].
+The following example was originally presented in [@Skjelvareid2012c], and is included here to give a more visual and intuitive sense of how `Synaptus` can create a focused image from raw ultrasonic data. For the sake of brevity, only a short description is given here, but the full details of the experiment can be found in [@Skjelvareid2012b].
 
-Two test blocks, one made of acrylic glass and the other of aluminium, were manufactured for the experiment. Each block had four flat-bottom holes
-with 3 mm diameter, spaced 20 mm apart. Such holes constitute point-like scatterers for ultrasonic imaging, and enable direct evaluation of the performance of focusing algorithms. The blocks were immersed in water, and the imaging geometry as a whole thus consisted of three layers with distinct sound wave velocities; 1480 m/s, 2730 m/s and 6320 m/s for water, acrylic glass and aluminium, respectively. The test blocks were imaged using a single 2.25 MHZ, 6 mm transducer, with a 1 mm spatial sampling interval. The imaging geometry is shown in Figure 1 and 2.
+Test blocks made of acrylic glass and aluminium were manufactured for the experiment. Each block had four flat-bottom holes
+with 3 mm diameter, spaced 20 mm apart. Such holes constitute point-like scatterers during ultrasonic imaging, and enable direct evaluation of the performance of focusing algorithms. The blocks were immersed in water, and the imaging geometry as a whole thus consisted of three layers with distinct sound wave velocities; 1480 m/s, 2730 m/s and 6320 m/s for water, acrylic glass and aluminium, respectively. The test blocks were imaged using a single 2.25 MHZ, 6 mm transducer, with a 1 mm spatial sampling interval. The imaging geometry is shown in Figure 1 and 2.
 
-The raw data from the ultrasonic scan is shown in Figure 3, with isosurfaces generated based on the ultrasonic responses from the flat-bottom holes. The image shows how the response from each hole is fairly broad, especially in the lowest layer. This effect is due to the width of the ultrasonic beam increasing with depth.
+The raw data from the ultrasonic scan is shown in Figure 3, with isosurfaces generated based on the responses from the flat-bottom holes. The image shows how the response from each hole is fairly broad, especially in the lowest layer. This effect is due to the width of the ultrasonic beam increasing with depth.
 
[…]
