### 2021-09-13 #3579 @mh-skjelvareid
@Kevin-Mattheus-Moerman - thanks for your recommendations! I've followed up on them, with comments for each issue (below).

> Here are some recommendations:
>
> * [x]  Change the `_readme.txt` file to a `README.md` markdown file. This way GitHub will recognize it and render it on the project front page. This helps folks clearly see what the project is about when they land on your project page.

The README file has been updated to a markdown file.

> * [x]  Change the `_license.txt` file to a plain `LICENSE` file. This way GitHub will recognize it and render a license badge at the top of your repository. This helps folks instantly see the license type for the project when they land on your page.

The LICENSE file has been updated to a markdown file.

> * [x]  Change the `_contributing.txt` file to a formatted `CONTRIBUTING.md` file and link to it from the readme.

The CONTRIBUTING file has been updated to a markdown file, and it is linked to from README.

> * [x]  Furthermore, in relation to your contributor community we recommend (not required) one adds a `COC.md` (code of conduct) too. Many developers use this template: https://www.contributor-covenant.org/version/2/1/code_of_conduct/code_of_conduct.md (do not forget to amend the contact details in the template if you use it).

A COC file with updated contact info has been added.

> * [x]  If you have  a couple of example images that help show the functionality I recommend you render these in your readme file as well as in the paper. You can also consider creating a "graphical abstract" figure that summarizes functionality. This will help folks to visually grasp what the code offers.

I've added a section called "Example raw and focused images" to the README file, with several visual examples of experiment setups, and "before and after" ultrasound data. These images have been taken directly from the PhD thesis distributed with the code, and include figure numbers and captions. This may look a bit odd, but I feel that having a direct reference to the thesis (with more in-depth descriptions of the examples) could be a good thing. However, I'm open to removing the captions if they seem out of place.

I've also included one of the examples in the paper itself. When I originally read the example paper for JOSS, I got the impression that the paper only needed "Summary" and "Statement of need" sections, and didn't think to include example figures. I will let you / the reviewers decide if you think the "Example use" section should be included in the paper or if it seems out of place. The information is included in the README file anyway, and I wasn't sure if I should duplicate it in the paper.

I haven't created a graphical abstract, but I think the example(s) are quite visual, and I think they may be enough in themselves.

> As an example, here is a link to one of my projects: https://github.com/gibbonCode/GIBBON, note the use of the `README.md`, the way a license badge is rendered, the inclusion of a graphical overview image, the linking to contributing guidelines, and a code of conduct. You may also like the way I render "badges" in the readme for the license etc. you can copy the syntax from the readme if you want to use this too.

The README file/page for GIBBON both looks nice and is very well structured - thanks for pointing me to it, it's a good example to follow. I've copied the use of a license badge in the README file.

>
> Quick comments on paper (we may have more later):
>
> * [x]  In `“raw data,”` it looks like the comma should be outside of the quotation marks e.g. `“raw data”, `
> * [x]  Check typo: `ultrasund` -> `ultrasound`

Fixed.
