### 2021-09-14 #3579 @Kevin-Mattheus-Moerman
@mh-skjelvareid thanks. Can you also work on that potentially missing DOI? Also it looks like the `“raw data,”` issue is still in there.
