+++ paper/paper.bib
@@ -142,9 +142,9 @@
   doi = {10.1016/j.ndteint.2012.10.005}
 }
 
-@misc{NSGeophysics2017,
+@misc{stolt_fk_mig,
   author = {Alain Plattner},
-  title = {Near Surface Geophysics},
+  title = {stolt_fk_mig},
   year = {2017},
   publisher = {GitHub},
   journal = {GitHub repository},
