+++ paper/paper.bib
@@ -201,7 +201,7 @@
 }
 
 @misc{beyond_epica,
-  author = {Beyond EPICA team},
+  author = {Carlo Barbante},
   title = {Beyond EPICA - Oldest ice},
   year = {2021},
   journal = {Beyond EPICA website},
