### 2021-09-14 #3579 @mh-skjelvareid
@Kevin-Mattheus-Moerman , I've fixed the missing DOI (the reference was to a Github repository that doesn't have a DOI, but the name of the repository is similar to another publication).

That `"raw data,"` issue is a real mystery, however. If you check the [markdown file for the paper](https://github.com/mh-skjelvareid/synaptus/blob/master/paper/paper.md), it's correct (i.e., `"raw data",`) while in the compiled PDF it changes to `"raw data,"` . A LaTeX issue?
