### 2021-09-15 #3579 @arfon
Wow, that's deeply peculiar! No idea. I even tried writing this out as `&quot;raw data&quot;&#44;` in the `paper.md` but to no avail. I suggest we don't get blocked on this for now, but perhaps @tarleb has some ideas?
