### 2021-09-15 #3579 @tarleb
Could you try setting

    ---
    lang: en-GB
    ---

in the YAML metadata block to force the comma out of the quote?
