### 2021-09-15 #3579 @arfon
@mh-skjelvareid – could you merge this PR so we can try @tarleb's suggestion? https://github.com/mh-skjelvareid/synaptus/pull/4
