+++ paper/paper.md
@@ -15,7 +15,6 @@ affiliations:
    index: 1
 date: 4 August 2021
 bibliography: paper.bib
-lang: en-GB
 ---
 
 **NOTE:** This is a _draft_ version of a paper submitted to the Journal of Open Source Software (JOSS).
@@ -27,7 +26,7 @@ Ultrasonic imaging is performed by transmitting sound waves into an object and r
 
 Transducer elements can be stacked along a line or on a grid to create transducer *arrays*. By manipulating how signals are transmitted by each individual element, different shaped waves can be created, e.g. a plane wave, or a wave focused at a point. The spatial distribution of the elements also enables recording of a 2D or 3D map of reflections without moving the array as a whole.
 
-The recordings of scattered waves represent "raw data", both for single-transducer and array imaging setups. This data often suffers from poor resolution, making it hard (if not impossible) to interpret directly. The data needs to be focused in order to create an image resembling the physical structure of an object. This process - creating focused images from raw pulse-echo ultrasound data - is the purpose of the `Synaptus` toolbox.
+The recordings of scattered waves represent raw data, both for single-transducer and array imaging setups. This data often suffers from poor resolution, making it hard (if not impossible) to interpret directly. The data needs to be focused in order to create an image resembling the physical structure of an object. This process - creating focused images from raw pulse-echo ultrasound data - is the purpose of the `Synaptus` toolbox.
 
 Sonic imaging is used in many different applications, such as geophysical imaging, sonar imaging, medical imaging, and non-destructive testing (NDT) of industrial components. The methods and algorithms used in these fields are similar in many ways, but there are also significant differences in hardware, scale, and physical properties of the objects that are imaged. The software presented here was originally written as part of a PhD thesis on **ultrasound imaging for non-destructive testing** [@Skjelvareid2012b].
 
