+++ experimental/stolt_mig_2d/stolt_mig_2d.m
@@ -1,0 +1,255 @@
+function varargout = stolt_mig_2d(ptx,fs,tDelay,cc,fLow,fHigh,xStep,varargin)
+% MULOK - Multi-Layer Omega-K algorithm for synthetic aperture ultrasound
+%
+%   Usage:
+%   ...
+%
+%   Input parameters:
+%   ptx         -   2D pulse-echo data
+%   tDelay      -   Time delay between pulse transmission and measurement [s]
+%   cc          -   Vector of comp. wave velocities for each layer [m/s]
+%   thick       -   Vector of thickness for each layer [m/s]
+%                   (set last boundary equal to end of ROI)
+%   fLow        -   Lower limit of transducer frequency band [Hz]
+%   fHigh       -   Higher limit of transducer frequency band [Hz]
+%   xStep       -   spatial step size(s) (distance between measurements) [m]
+%
+%   Optional parameter-value input pairs:
+%   tFftMult    -   multiplier for t-axis FFT size. Default: 1
+%   xFftMult    -   multiplier for x-axis FFT size. Default: 1
+%   yFftMult    -   multiplier for y-axis FFT size. Default: 1
+%   upSamp      -   omega upsamp. factor (preproc. before interp.) Default: 4
+%                   Used only for linear interpolation
+%
+%   Output parameters (varargout):
+%   im          -   Reconstructed image as cell array - one cell per layer
+%   xIm         -   x positions for pixels in im
+%
+%
+%   LICENSE DISCLAIMER:
+%   This program is free software: you can redistribute it and/or modify
+%   it under the terms of the GNU General Public License as published by
+%   the Free Software Foundation, either version 3 of the License, or
+%   (at your option) any later version.
+%
+%   This program is distributed in the hope that it will be useful,
+%   but WITHOUT ANY WARRANTY; without even the implied warranty of
+%   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+%   GNU General Public License for more details.
+%
+%   You should have received a copy of the GNU General Public License
+%   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+%
+%   2021-09-14  MHS
+%
+%   Copyright (C) 2021  Martin H. Skjelvareid
+
+
+%% Parse optional input parameters
+% p = inputParser;                        % Input parsing object
+% p.addParameter('xFftMult',1);           % Multiplier for FFT size in x dir.
+% p.addParameter('yFftMult',1);           % Multiplier for FFT size in y dir.
+% p.addParameter('tFftMult',1);           % Multiplier for FFT size in t dir.
+% p.addParameter('zFftMult',1)            % Multiplier for FFT size in z dir.
+% p.addParameter('upSamp',4)              % Multiplier for interpolation FFT size
+% p.addParameter('hh',1);                 % Impulse response / waveform
+% p.addParameter('xStart',0)              % First x value of scan
+% p.addParameter('yStart',0)              % First x value of scan
+% validator = @(str) any(strcmp(str,{'linear','chirpz'}));
+% p.addParameter('interpol','linear',validator);     % Interpolation method
[…]
