+++ experimental/stolt_mig_2d/mlok2d.m
@@ -1,0 +1,203 @@
+function varargout = mlok2d(ptx,fs,mwb,cc,thick,fLow,fHigh,xStep,varargin)
+%
+%	Usage:
+%	[im,...] = mlok2d(ptx,fs,mwb,cc,thick,fLow,fHigh,xStep,...)
+% 
+%	Input parameters:
+%	ptx			-	p(t,x): matrix with ultrasonic data, A-scans along columns
+%	fs			-	Sampling frequency [Hz]
+%	mwb			-	Measurement Window Begin [s] - time delay between
+%					pulse transmission and beginning of data logging
+%	cc			-	Vector of compressional wave velocities for each layer [m/s]
+%	thick		-	Vector of thickness for each layer [m/s] (set last boundary equal to end of ROI)
+%	fLow		-	Lower limit of transducer frequency band [Hz]
+%	fHigh		-	Higher limit of transducer frequency band [Hz]
+%	xStep		-	x step value (distance between measurements) [m]
+%
+%	Optional parameter-value input pairs:
+%	xFftMult	-	multiplier for FFT size in x dir. Default: 1
+%	tFftMult	-	multiplier for FFT size in t dir. Default: 1
+%	zFftMult	-	multiplier for z FFT size. Default: 1 
+%   upSamp      -   omega upsampling factor (preproc. before interp.) Default: 4
+%	hh			-	Transducer impulse response, sampled at fs. Used in matched filter. Default: 1 (no filt.)
+%	xStart		-	First x value of scan [m]. Default: 0. Only relevant for "xIm" output
+%
+%	Output parameters (varargout):
+%	im			-	Reconstructed image as cell array - one cell per layer
+%	xIm			-	x positions for pixels in im (single vector)
+%	zIm			-	z positions for pixels in im (cell array, different z res. for each layer)
+%
+%   Note: The z resolution of the output image for layer l is set to 
+%       dz = (cc(l)/2)/fHigh. 
+%   This is done to minimize the required number of z depths for each
+%   layer. If this resolution is too coarse, try resampling the image to a
+%   higher resolution (see resample, interp)
+%
+%	2010-06-16	Martin H. Skjelvareid
+%
+%   2010-10-20 Added a term effectively doubling nFFTz. It is assumed that
+%              this will correct aliasing errors seen earlier. 
+
+%	A note on code notation:
+%	------------------------
+%	What Stolt migration does is essentially to transform the sampled
+%	wavefield p(t,z=0,x) to p(t=0,z,x). To understand how this is done,
+%	consult f.ex. Margrave (2003) - "Numerical Methods in Exploration
+%	Seismology".
+%
+%	In the process of doing this transformation, many versions of the
+%	wavefield are used. A short explanation of the code notation is
+%	included here to make reading of the code easier.
+%
+%	Code variable			Represents wave field
+%	ptx						p(t,x)
+%	Pox						P(omega,x)
+%	Pokx					P(omega,k_x)
+%	Pkzkx					P(k_z,k_x)
+%	pzx						p(z,x)
+
+
+%% Parse optional input parameters
+p = inputParser;									% Input parsing object
+p.addParamValue('xFftMult',1);						% Multiplier for FFT size in x dir.
+p.addParamValue('tFftMult',1);						% Multiplier for FFT size in t dir.
[…]

[…]
