### 2021-09-16 #3579 @mh-skjelvareid
Thanks for trying, @arfon ! I've reverted the change, and also decided to remove the quotation marks altogether - they're not really required in the context.
