### 2021-12-14 #3579 @kylemath
sorry I wll not be able to help with this paper due to work constraints but
I hope to help with these in the future

On Tue, Dec 7, 2021 at 3:03 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> @michaelberks <https://github.com/michaelberks> @kylemath
> <https://github.com/kylemath> @ajmalkoti <https://github.com/ajmalkoti>
> would you be interested in reviewing this work (a MATLAB/octave toolbox for
> ultrasound imaging) for the Journal of Open Source Software (JOSS)
> <https://joss.theoj.org>?
> The smooth review process
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> takes
> place on GitHub, and focuses on the software
> <https://github.com/mh-skjelvareid/synaptus> as well as a short paper
> <https://github.com/openjournals/joss-papers/blob/joss.03579/joss.03579/10.21105.joss.03579.pdf>
> .
>
> Thanks!
>
> Note under the current COVID19 pandemic we can be quite flexible in terms
> of giving your time to complete the review. If you are able to review this
> work in principle but currently lack the time to do so, feel free to
> indicate a later time when you may or estimate how long it would take to
> complete.
>
> Note this is a pre-review issue in which reviewers are assigned. Once
> sufficient reviewers are recruited a dedicated review issue is initiated.
> In that review issue the reviewers are guided using check lists. Here is an
> example of a recently completed review issue to give you an idea of what it
> is like to review for JOSS: #2548
> <https://github.com/openjournals/joss-reviews/issues/2548>
>
