### 2022-02-16 #3579 @CPBridge
@Kevin-Mattheus-Moerman  while I'm happy to help out where my expertise matches, I'm afraid I think this is too far outside my domain of knowledge to be able to provide a useful review. I have no experience with ultrasound image formation or raw signal processing or the algorithms implemented by this toolbox, only image analysis. Apologies.
