### 2022-02-18 #3579 @emanuelhuber
Hello
@Kevin-Mattheus-Moerman  Thank you very much for asking!
I don't have a Matlab license and I don't want to install Octave on my computer. So I will not test this toolbox. But at could review the paper, the code (I know Matlab/octave) and the document. I would need 1 week time. Let me know if you need me.
Best,
Emanuel
