### 2022-02-22 #4185 @whedon
[REVIEW]: Synaptus: A Matlab/Octave toolbox for synthetic aperture ultrasound imaging
**Submitting author:** @mh-skjelvareid (<a href="http://orcid.org/0000-0003-2103-865X">Martin H. Skjelvareid</a>)
**Repository:** https://github.com/mh-skjelvareid/synaptus
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
