### 2022-03-13 #4185 @emanuelhuber
(@Kevin-Mattheus-Moerman should I leave the check box "Installation instructions" unchecked if the requirement seems to be not completely fulfilled?

Here are some points I rise

## Documentation

- readme.md file: In the summary the author explains that the toolbox is for a certain type of data but he does not explain what exactly the toolbox do (i.e., focusing these kind of data). I do not know if the toolbox can do more than focusing.
- There is no "clearly-stated list of dependencies": I had to install the 'signal' toolbox which required the 'control' toolbox. A short information about the dependence on signal and how to install and load the 'signal' toolbox would be welcome (why should I load manually the signal package?).
- I could not find any "external" API method documentation but the function are documented in their file and the three main focusing methods are described in the readme file and the examples allow to find the corresponding functions. It requires some extra "dig work" for novices to octave (like me). The simplified algorithms are surely a great ressource for the community.
- examples:
	- Both file 'array_psm_testFullMatrix_1layer.m' and 'array_psm_testFullMatrix_2layer.m' have exactly the same note ('copper block') but load different files ('steel pins' and 'copper block'). Please check if the notes are correct.
- When running 'cpsm_test2D.m' I got the following error :
   	```
	All image ranges processed
	Total processing time: 1.3711 seconds.
	error: text: Invalid combination of points and text strings
	error: called from
	    text at line 128 column 7
	    polRawDataImage at line 93 column 1
	    cpsm_test2D at line 30 column 1
    	```

## Software paper

- Statement of need:
	- I suggest the author to repeat that synaptus is a toolbox for focusing...
	- The author writes that Alain Plattner used "the code". This is misleading and it would be better to replace "the code" with the "Stolt fk migration function".
	- I found other open-source toolboxes for SAR data focusing, it would be interesting to know how they differ from synaptus:
		- https://github.com/ThatcherC/RadarPro
		- https://github.com/cryotools/osaris
		- http://step.esa.int/main/toolboxes/snap/snap-faq/
	- Just as a remark: yes, Python is very popular and [Julia](https://julialang.org/) seems even more promising.

## Conclusion

The paper is well written, the code is documented, datasets are available, the examples work well, I could run everything although I am not an octave/matlab user and some effort has been done to explains some of the algorithm. I warmly thank the author for his contribution to the community!
