### 2022-03-16 #4185 @mh-skjelvareid
@emanuelhuber , thank you for your valuable feedback! I will try to follow up on and correct each point that you have raised. Thanks also for being patient and flexible regarding installation of octave. As indicated in the paper, I'd like to update the toolbox to an open source (and perhaps more modern) programming language, for example Python. I haven'y used Julia but I will definitely look into it.

@Kevin-Mattheus-Moerman : I guess I should wait for Samuel-Wagner to finish his review before I start changing anything?
