### 2022-03-16 #4185 @emanuelhuber
@mh-skjelvareid , I startet programming with Matlab 20 years ago, then I switched to R (which is very similar to Matlab). Now I am learning Python for my job and I don't like it. I wrote a R-package (RGPR) and am still playing with the thought to translate it in another language. First I thought that Python would be great because many people in geosciences use it. But now I don't know... maybe write C/C++ functions that can be used in any language? Write a Julia package?
