### 2022-03-16 #4185 @Kevin-Mattheus-Moerman
@emanuelhuber
> ... should I leave the check box "Installation instructions" unchecked if the requirement seems to be not completely fulfilled?

Yes, leave anything you are not happy with unchecked until resolved.
