+++ paper/paper.bib
@@ -207,3 +207,27 @@
   journal = {Beyond EPICA website},
   url = {https://www.beyondepica.eu/en/}
 }
+
+@article{osaris,
+    author={Loibl, David and Bookhagen, Bodo and Valade, Sébastien and Schneider, Christoph},
+    title={OSARIS, the “Open Source SAR Investigation System” for Automatized Parallel InSAR Processing of Sentinel-1 Time Series Data With Special Emphasis on Cryosphere Applications},
+    journal={Frontiers in Earth Science},
+    volume={7},
+    year={2019},
+    url={https://www.frontiersin.org/article/10.3389/feart.2019.00172},
+    doi={10.3389/feart.2019.00172}
+}
+
+@inproceedings{snap,
+       author = {{Zuhlke}, Marco and {Fomferra}, Norman and {Brockmann}, Carsten and {Peters}, Marco and {Veci}, Luis and {Malik}, Julien and {Regner}, Peter},
+        title = "{SNAP (Sentinel Application Platform) and the ESA Sentinel 3 Toolbox}",
+    booktitle = {Sentinel-3 for Science Workshop},
+         year = 2015,
+       editor = {{Ouwehand}, L.},
+       series = {ESA Special Publication},
+       volume = {734},
+        month = dec,
+          eid = {21},
+        pages = {21},
+          url = {https://ui.adsabs.harvard.edu/abs/2015ESASP.734E..21Z}
+}
