### 2022-03-26 a460044
Update README after comments from Emanuel Huber
