+++ README.md
@@ -1,8 +1,11 @@
 [![License](https://img.shields.io/badge/License-GNU_GPLv3-orange.svg)](https://github.com/mh-skjelvareid/synaptus/blob/master/LICENSE.md)
 
 # Summary
-Synaptus is a Matlab/Octave toolbox for synthetic aperture or array imaging. It was originally developed for ultrasonic imaging for non-destructive testing, but can be applied for similar imaging modes (e.g. ground penetrating radar). The toolbox focuses on algorithms implemented in the Fourier domain and on multilayered structures (e.g. water, metal, rock).
+Synaptus is a Matlab/Octave toolbox for synthetic aperture or array imaging. It was originally developed for ultrasonic imaging for non-destructive testing, but can be applied for similar imaging modes (e.g. ground penetrating radar). The toolbox focuses on algorithms implemented in the Fourier domain, and on imaging in multilayered structures (e.g. water, metal, rock).
 
+The core functionality of the toolbox is to create focused images from raw (unfocused) pulse-echo data. Such data is produced by a transducer that transmits waves into a propagating medium, and records backscattered waves from within the medium. A backscattered "echo" is created when the waves interact with an object or layer with different physical properties than the propagating medium, e.g. a metal object in water. A measurement at a single point in space thus produces a 1-dimensional "depth profile". By moving the transducer laterally relative to the object under study, it is possible to create a 2- or 3-dimensional image of the object. A similar measurement can be performed using an array of multiple transducers. Due to the divergence of the transducer beams, the echoes from scattering objects are "smeared" laterally, making the images unfocused and hard to interpret. Examples of such images are given in the "Example raw and focused images" section below.
+
+The algorithms in the Synaptus toolbox take such raw, unfocused images as input and processes them to create focused images. Conceptually, this is done by treating the raw images as measurements of a wave field, and using the wave equation to manipulate the wave field into a focused image. The details of this process is described in the PhD thesis included in the toolbox.
 
 
 # Background
@@ -61,9 +64,31 @@ The toolbox is organized into the following folders:
 - 'docs' contains relevant documentation (PhD thesis ++)
 - 'experimental' contains "draft" code related to smaller concepts and ideas, including tilt compensation. The code is not fully polished/commented.
 
+# Requirements
+The toolbox requires a working base installation of Matlab or Octave, with some additional signal processing functions. Details for Matlab and Octave are given below.
+
+## Matlab
+Matlab is available for Linux, MacOS and Windows, and can be purchased from [Mathworks](https://mathworks.com/store/). Running Synaptus on Matlab requires:
+- Base installation of Matlab (has been tested with version 2021a, installed on Windows)
+- [Signal processing toolbox](https://se.mathworks.com/products/signal.html) (has been tested with version 8.6)
+
+## Octave
+GNU Octave is free and open source, and is available for both Linux, MacOS and Windows. See [GNU download and installation page](https://www.gnu.org/software/octave/download) for details.
+
+Running Synaptus on Octave requires:
+- Base installation of Octave (has been tested with version 5.2.0, installed on Ubuntu 20.04). On Debian / Ubuntu systems, Octave can be installed with `sudo apt-get install octave`.
+- The [signal package](https://octave.sourceforge.io/signal/index.html) (has been tested with version 1.4.1). Octave installers for Windows bundle the signal package with the main application. On Debian/Ubuntu systems, the package can be installed using `sudo apt-get install octave-signal` Note that the signal package needs to be "loaded" inside Octave using the command `pkg load signal`. This can be automated by adding the command to a [Octave startup file](https://octave.org/doc/v6.4.0/Startup-Files.html#Startup-Files).
+
 
 # Installation
[…]
