### 2022-03-26 #4185 @mh-skjelvareid
@Kevin-Mattheus-Moerman , I have now tried to address the points raised by @emanuelhuber :

> **Documentation**
> * readme.md file: In the summary the author explains that the toolbox is for a certain type of data but he does not explain what exactly the toolbox do (i.e., focusing these kind of data). I do not know if the toolbox can do more than focusing.

I have updateed the summary to clarify that the only purpose of the toolbox is to focus pulse-echo data.

> * There is no "clearly-stated list of dependencies": I had to install the 'signal' toolbox which required the 'control' toolbox. A short information about the dependence on signal and how to install and load the 'signal' toolbox would be welcome (why should I load manually the signal package?).

I have added "Requirements" section to the readme.md file which states that the toolbox requires Matlab / Octave and a signal toolbox to run. I've added a few practical details regarding installation of these.

> * I could not find any "external" API method documentation but the function are documented in their file and the three main focusing methods are described in the readme file and the examples allow to find the corresponding functions. It requires some extra "dig work" for novices to octave (like me). The simplified algorithms are surely a great ressource for the community.

I've added a "Documentation" section which states that the main toolbox functions are documented in the standard Matlab/Octave style (description at start of function). The documentation can be viewed on the Matlab/Octave command line by using the `help` command. I've also described how the "test" scripts are intended as a demonstration of how to use the toolbox.

> * Both file 'array_psm_testFullMatrix_1layer.m' and 'array_psm_testFullMatrix_2layer.m' have exactly the same note ('copper block') but load different files ('steel pins' and 'copper block'). Please check if the notes are correct.

The description in 'array_psm_testFullMatrix_1layer.m' was incorrect and has been updated - thanks!

> * When running 'cpsm_test2D.m' I got the following error :
> All image ranges processed
> Total processing time: 1.3711 seconds.
> error: text: Invalid combination of points and text strings
> error: called from
>     text at line 128 column 7
>     polRawDataImage at line 93 column 1
>     cpsm_test2D at line 30 column 1
> 	```

The `polRawDataImage` and `polFocusedImage` functions have been fixed to avoid this error.

> **Software paper**
> Statement of need:
> * I suggest the author to repeat that synaptus is a toolbox for focusing...

I wasn't entirely sure what Emanuel meant by this comment, but I'm guessing he wanted a similar clarification
as in the readme.md file (a statement that the toolbox is only used for creating focused images). I have tried to add this in the introduction to the Statement of need.

> * The author writes that Alain Plattner used "the code". This is misleading and it would be better to replace "the code" with the "Stolt fk migration function".

This passage has been updated as recommended.

> * I found other open-source toolboxes for SAR data focusing, it would be interesting to know how they differ from synaptus:
> https://github.com/ThatcherC/RadarPro
> https://github.com/cryotools/osaris
> http://step.esa.int/main/toolboxes/snap/snap-faq/

I have added a paragraph to the paper referencing some of these toolboxes, and explaining that
* They are generally so specialized towards radar data that they are hard to use for ultrasonic data
* They cannot process multi-layer data, which is one of the main features of the `Synaptus` toolbox.
