### 2022-03-29 #4185 @emanuelhuber
@Kevin-Mattheus-Moerman All check boxes are now checked!
