### 2022-07-28 16b0977
Fixed test files in relation to path names on Ubuntu linux
