### 2022-07-28 #4185 @Kevin-Mattheus-Moerman
@arfon given that it has been challenging to find reviewers for this submission and that one of the reviewers was unable to complete the review I proposed that I will review this work as a second reviewer. Given my MATLAB and image analysis skills I feel I can review this quite well.
