### 2022-07-28 #4185 @Kevin-Mattheus-Moerman
@mh-skjelvareid I installed and tested the software and it all works really well after I implemented the below minor fix.

# Main issue

- [ ] Can you merge [this pull request](https://github.com/mh-skjelvareid/synaptus/pull/5), it fixes a bug for Ubuntu Linux. Namely that the path definitions did not work. I went through the various test codes and replaced stuff like:
```matlab
%% Add path to necessary functions
addpath('../core','../misc')

%% Load test data
load('../datasets/ArraySteelPins.mat','data','MWB')
```

with
```matlab

%% Add path to necessary functions
toolboxPath=fileparts(fileparts(mfilename('fullpath'))); %Get the toolbox path

%Add core and misc path
addpath(fullfile(toolboxPath,'core'),fullfile(toolboxPath,'misc'));

%% Load test data
load(fullfile(toolboxPath,'datasets','ArraySteelPins.mat'),'data','MWB')
```
This fix should work for all operational systems.

# Minor comments (not requirements):
- [ ] Note that you currently use both `caxis([minVal maxVal])` as well as `set(gca,'CLim',[minVal maxVal])`, for clarity you could consider just using one approach. The former (the use of `axis`) is labelled as deprecated for future MATLAB releases in favor of `clim([minVal maxVal])`. However, do check compatibility with Octave too which may still use the `axis` approach.
- [ ] Note that `clear all` is the same as `clear` since `all` is default. Elsewhere you use `clearvars` which probably does the same thing again. But again, check Octave compatibility too.
- [ ] You sometime have `disp('Loading data');` and `disp('Processing data');` as messages in the tests, but not always, consider using it for all tests.
- [ ] You may consider, in future work, implementing automated testing, e.g. by running all tests automatically (see also something like: https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m).
- [ ] You may also consider, in future work, making the tests "publishable" in the MATLAB code publishing sense, this way they can turn in to HTML files that you can render as documentation too. Once again GIBBON might be an example.
- [ ] Similarly you may consider automated script based installation (which could simple be adding the toolbox paths and saving them).
