### 2022-08-01 85ad6cf
Fixed progress update (layer number) bug.
