### 2022-08-01 4a2e0df
Changes after review by Kevin Mattheus Moerman
Small changes regarding consistency in use of
- clearvars (rather than clear all)
- use of caxis([minVal, maxVal])
- Displaying "Loading data" and "Processing data"
- Displaying processing time for each test

In addition, a script for running all tests has been added, and each test has been given a short "header" that is printed out to the command line, so that it's possible to follow which test is being run.
