+++ test/cpsm_test3D.m
@@ -1,6 +1,12 @@
 close all
 clearvars
-clc
+
+%% Show test header (useful when run as part of test suite)
+disp('-----------------------------------------------------------------')
+disp('-----        SYNAPTUS TOOLBOX TEST: cpsm_test3D             -----')
+disp('-----    Cylindrical phase shift migration, 3D dataset      -----')
+disp('-----------------------------------------------------------------')
+disp(' ')
 
 %% Add path to necessary functions
 toolboxPath=fileparts(fileparts(mfilename('fullpath'))); %Get the toolbox path
@@ -9,6 +15,7 @@ toolboxPath=fileparts(fileparts(mfilename('fullpath'))); %Get the toolbox path
 addpath(fullfile(toolboxPath,'core'),fullfile(toolboxPath,'misc'));
 
 %% Load data and convert to double precision
+disp('Loading data')
 load(fullfile(toolboxPath,'datasets','CylScan3D_LeadSpheres.mat'),'ptpz','fs','phiStep','zStep',...
     'tDelay','r0','fLow','fHigh','cc');
 ptpz = double(ptpz);
@@ -21,10 +28,11 @@ rEnd = rStart + (size(ptpz,1)/fs)*(cc/2);
 transFunc = 'haun';     % Transfer function, options 'haun', 'gardner', 'exact'
 
 %% Focus image
-tic;
+disp('Processing data')
+tic
 [im,phiIm,rIm,zIm] = cpsm(ptpz,fs,tDelay,cc,fLow,fHigh,[phiStep zStep],...
     r0,rStep,rStart,rEnd,'transFunc',transFunc);
-disp(['Total processing time: ' num2str(toc) ' seconds.'])
+toc 
 
 %% Raw data envelope
 envRaw = zeros(size(ptpz));
@@ -45,3 +53,6 @@ imagesc(zIm*1e3,phiIm*(180/pi),abs(squeeze(max(im))));
 xlabel('z [mm]')
 ylabel('phi [deg]')
 title('Focused image C-scan (max. envelope)')
+
+%% Add blank line (nicer formatting for test text output).
+disp(' ')
