+++ test/cpsm_test2D.m
@@ -1,6 +1,12 @@
 close all
 clearvars
-clc
+
+%% Show test header (useful when run as part of test suite)
+disp('-----------------------------------------------------------------')
+disp('-----        SYNAPTUS TOOLBOX TEST: cpsm_test2D             -----')
+disp('-----    Cylindrical phase shift migration, 2D dataset      -----')
+disp('-----------------------------------------------------------------')
+disp(' ')
 
 %% Add path to necessary functions
 toolboxPath=fileparts(fileparts(mfilename('fullpath'))); %Get the toolbox path
@@ -9,6 +15,7 @@ toolboxPath=fileparts(fileparts(mfilename('fullpath'))); %Get the toolbox path
 addpath(fullfile(toolboxPath,'core'),fullfile(toolboxPath,'misc'));
 
 %% Load data, extract 2D slice and convert to double precision
+disp('Loading data')
 load(fullfile(toolboxPath,'datasets','CylScan3D_LeadSpheres.mat'),'ptpz','fs','phiStep','zStep',...
     'tDelay','r0','fLow','fHigh','cc');
 zSliceInd = 30;                         % Index for 2D slice from 3D dataset
@@ -23,10 +30,11 @@ rEnd = rStart + (size(ptp,1)/fs)*(cc/2);    % End of focused image
 transFunc = 'haun';     % Transfer function, options 'haun', 'gardner', 'exact'
 
 %% Focus
-tic;
+disp('Processing data')
+tic
 [im,phiIm,rIm] = cpsm(ptp,fs,tDelay,cc,fLow,fHigh,phiStep,r0,rStep,...
     rStart,rEnd,'transFunc',transFunc);
-disp(['Total processing time: ' num2str(toc) ' seconds.'])
+toc
 
 %% Plot raw data and focused image
 figure
@@ -34,3 +42,6 @@ polRawDataImage(abs(hilbert(ptp)),tDelay+(0:(size(ptp,1)-1))/fs,phiIm,...
     [],[],r0,cc)
 figure
 polFocusedImage(abs(im),phiIm,rIm,[],[])
+
+%% Add blank line (nicer formatting for test text output).
+disp(' ')

[…]
