### 2022-08-01 4f9770d
Added mention of "tests_runAll" script under installation instructions.
