### 2022-08-01 #4185 @mh-skjelvareid
@Kevin-Mattheus-Moerman  Thanks a lot for the nice fix for the path definitions, and the other comments regarding consistency and testing. I've addressed the main point and some of the listed suggestions. I've checked the boxes for issues that have been fixed, and have added comments in between.

> @mh-skjelvareid I installed and tested the software and it all works really well after I implemented the below minor fix.
>
> # Main issue
> * [x ]  Can you merge [this pull request](https://github.com/mh-skjelvareid/synaptus/pull/5), it fixes a bug for Ubuntu Linux. Namely that the path definitions did not work. I went through the various test codes and replaced stuff like:
>
> ```matlab
> %% Add path to necessary functions
> addpath('../core','../misc')
>
> %% Load test data
> load('../datasets/ArraySteelPins.mat','data','MWB')
> ```
>
> with
>
> ```matlab
> %% Add path to necessary functions
> toolboxPath=fileparts(fileparts(mfilename('fullpath'))); %Get the toolbox path
>
> %Add core and misc path
> addpath(fullfile(toolboxPath,'core'),fullfile(toolboxPath,'misc'));
>
> %% Load test data
> load(fullfile(toolboxPath,'datasets','ArraySteelPins.mat'),'data','MWB')
> ```
>
> This fix should work for all operational systems.

Done!

>
> # Minor comments (not requirements):
> * [x ]  Note that you currently use both `caxis([minVal maxVal])` as well as `set(gca,'CLim',[minVal maxVal])`, for clarity you could consider just using one approach. The former (the use of `axis`) is labelled as deprecated for future MATLAB releases in favor of `clim([minVal maxVal])`. However, do check compatibility with Octave too which may still use the `axis` approach.

I've gone through the code and changed this to consistent use of caxis([minVal,maxVal]).

> * [x ]  Note that `clear all` is the same as `clear` since `all` is default. Elsewhere you use `clearvars` which probably does the same thing again. But again, check Octave compatibility too.

I've gone through and changed the first lines of each script to "close all; clearvars". "clearvars" only clears variables while "clear" clears everything. It doesn't matter much for the test scripts, I think, but I've found clearvars to work better in some cases, so I've made it a habit.

> * [x ]  You sometime have `disp('Loading data');` and `disp('Processing data');` as messages in the tests, but not always, consider using it for all tests.

I've added this for all the tests, and I've also standardized the measurement of processing time for each test (using only tic and toc).

> * [x ]  You may consider, in future work, implementing automated testing, e.g. by running all tests automatically (see also something like: https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m).

I've added a new script, "tests_runAll.m", which displays a "test suite header" and then runs each test one by one. Inside each of the tests I've also added a bit of code which displays a header. This gives a nice separation between the text output ("Processing layer 1 of 2" etc.) from each test.

> * [ ]  You may also consider, in future work, making the tests "publishable" in the MATLAB code publishing sense, this way they can turn in to HTML files that you can render as documentation too. Once again GIBBON might be an example.

This seems like a nice solution. I'm not sure if I understand exactly how it works, though. Is all the documentation generated through (Matlab) publishing during installation? If so I guess it won't work for Octave? I get the sense that this could make for some nice documentation, but that it'll be a bit too much work for me to implement at this stage (the toolbox is no longer part of my active research work).

> * [ ]  Similarly you may consider automated script based installation (which could simple be adding the toolbox paths and saving them).

I agree that this would make the installation a bit smoother. However, I have no experience in making installation scripts, and I'm expecting that it's hard to make something that works in absolutely all cases...(?) I think I'd prefer for the users to add the toolbox to the path of their particular installation of Matlab / Octave themselves.
