### 2022-08-04 #4185 @Kevin-Mattheus-Moerman
@mh-skjelvareid thanks for implementing those changes and for your comments. I tested the changes this time using Octave on Ubuntu, rather than MATLAB, and all tests passed fine and the changes look good.
