### 2022-08-05 #4185 @Kevin-Mattheus-Moerman
@mh-skjelvareid I have proofread the paper and here are some last points and recommendations:

- [ ] Please complete the affiliation with the full name of the institute, the city name, and the country.
- [ ] Remove the statement: `NOTE: This is a draft version of a paper submitted to the Journal of Open Source Software (JOSS).`
- [ ] Check `more computationally efficient that operating` which should read `more computationally efficient than operating`
- [ ] Check `resouce` -> `resource`
- [ ] Check `cartesian` -> `Cartesian'
- [ ] Check `procesing` -> `processing`
- [ ] Check `accomodate` -> `accommodate`
- [ ] Can you check for consistency with either American or European (Brittish/Irisih) English? E.g. You use `optimized`, `vectorized`, and `specialized` which favors American English, whereas the use of Aluminium over Aluminum favours Brittish English. Either is fine but it is best to be consistent.
- [ ] Check `Overview of experiment setup`, this should perhaps be improved to something like `Overview of the experimental setup`
- [ ] Check `Seen from side` -> `Seen from the side`
- [ ] Check `ultrasonic raw data` -> `raw ultrasonic data`
