### 2022-08-08 32c2bed
Paper corrections after review by Kevin M. Moerman + spellchecking
