+++ paper/paper.md
@@ -11,13 +11,12 @@ authors:
     orcid: 0000-0003-2103-865X
     affiliation: 1
 affiliations:
- - name: UiT - the Arctic University of Norway
+ - name: Department of Computer Science and Computational Engineering, UiT - the Arctic University of Norway, Narvik, Norway
    index: 1
 date: 4 August 2021
 bibliography: paper.bib
 ---
 
-**NOTE:** This is a _draft_ version of a paper submitted to the Journal of Open Source Software (JOSS).
 
 # Summary
 `Synaptus` is a toolbox made for focusing ultrasound images. It can, however, also be used on data acquired using similar imaging principles, e.g. sonar and radar images.
@@ -32,9 +31,9 @@ Sonic imaging is used in many different applications, such as geophysical imagin
 
 The thesis focused on three main points:
 
-* Processing data in the Fourier domain (faster and sometimes simpler than procesing in the time-space domain)
+* Processing data in the Fourier domain (faster and sometimes simpler than processing in the time-space domain)
 * Adapting single-layer algorithms to multi-layered media, e.g. water and metal
-* Adapting algorithms originally made for cartesian coordinates to cylindrical coordinates (better suited for imaging pipes from the inside)
+* Adapting algorithms originally made for Cartesian coordinates to cylindrical coordinates (better suited for imaging pipes from the inside)
 
 The thesis is included in the repository and contains a full list of references. However, the most important ones will be highlighted here: `Synaptus` was greatly inspired by previous work by Tomas Olofsson and Tadeusz Stepinski at Uppsala University [@Olofsson2010, @Stepinski2007]. The underlying theory (mainly phase shift migration and Stolt migration) was originally developed for geophysical imaging [@Stolt1978; @Gazdag1978], and the free book and software on exploration seismology available through CREWES [@crewes_toolbox_2021] was also very helpful during software development. An extended (not free) version of the book is also available [@margrave_lamoureux_2019]. Finally, the book "Fourier Acoustics" by E.G. Williams [@Williams1999] was essential in developing the theory for cylindrical imaging geometries.
 
@@ -42,17 +41,17 @@ The thesis is included in the repository and contains a full list of references.
 
 # Statement of need
 
-`Synaptus` is a Matlab / Octave toolbox for creating focused images from raw ultrasonic data using the concept of synthetic aperture. The core algorithms have been written as a small set of functions that can handle many different types of datasets (2D / 3D data, single- or multilayered media, cartesian or cylindrical geometries). The code is highly vectorized, taking advantage of optimized libraries for linear algebra in Matlab / Octave. In addition, the algorithms operate in the Fourier domain, which in many cases is more computationally efficient that operating in the time-space domain ("delay-and-sum").
+`Synaptus` is a Matlab / Octave toolbox for creating focused images from raw ultrasonic data using the concept of synthetic aperture. The core algorithms have been written as a small set of functions that can handle many different types of datasets (2D / 3D data, single- or multilayered media, Cartesian or cylindrical geometries). The code is highly vectorized, taking advantage of optimized libraries for linear algebra in Matlab / Octave. In addition, the algorithms operate in the Fourier domain, which in many cases is more computationally efficient than operating in the time-space domain ("delay-and-sum").
 
-The toolbox includes a number of scripts that test the core algorithms by running them on a set of relevant datasets. The datasets represent a valuable resouce in themselves, given that there are very few publicly available NDT ultrasound datasets. The datasets can e.g. be used as benchmarks by researchers working on new algorithms.
[…]
