### 2022-08-08 501c5b8
Update to enable spectral input (not standard time-domain input)
