+++ experimental/tilt2d/test_tilt2d.m
@@ -23,7 +23,7 @@ bb = pos.z(1) - aa*pos.x(1);
 figure
 colormap(gray(256))
 imagesc(xPlot*1e3,zPlot*1e3,logImage(ptx))
-set(gca,'CLim',[-40 0])
+caxis([-40 0])
 xlabel('X pos. [mm]')
 ylabel('Z pos. ref. water [mm]')
 title('Original B-scan')
@@ -43,7 +43,7 @@ xPlotNew = (0:(nX-1))*xStepNew;
 figure
 colormap(gray(256))
 imagesc(xPlotNew*1e3,zPlot*1e3,logImage(real(ptxTilt)))
-set(gca,'CLim',[-40 0])
+caxis([-40 0])
 xlabel('X pos. [mm]')
 ylabel('Z pos. ref. water [mm]')
 title('B-scan extrapolated to tilted surface')
@@ -52,13 +52,13 @@ hold on
 %% Use wavefield at surface as starting point for focusing inside block
 tDelay_tilt = 0;
 thick_copper = 0.08;
-[im,xIm,zIm] = psm(ptxTilt,fs,tDelay_tilt,cc(2),thick_copper,fLow,fHigh,xStepNew);
+[im,xIm,zIm] = psm(real(ptxTilt),fs,tDelay_tilt,cc(2),thick_copper,fLow,fHigh,xStepNew);
 
 %% Show focused image
 figure
 colormap(gray(256))
 imagesc(xIm*1e3,zIm{1}*1e3,logImage(im{1}))
-set(gca,'CLim',[-40 0])
+caxis([-40 0])
 xlabel('X [mm]')
 ylabel('Z [mm]')
 title('Focused image of copper block')
