### 2022-08-08 #4185 @mh-skjelvareid
@Kevin-Mattheus-Moerman, thanks for finding all these mistakes, I've corrected them all, and have also corrected a couple of other mistakes I found during spellchecking. I went for consistency with American English.

> @mh-skjelvareid I have proofread the paper and here are some last points and recommendations:
>
> * [x]  Please complete the affiliation with the full name of the institute, the city name, and the country.
> * [x]  Remove the statement: `NOTE: This is a draft version of a paper submitted to the Journal of Open Source Software (JOSS).`
> * [x]  Check `more computationally efficient that operating` which should read `more computationally efficient than operating`
> * [x]  Check `resouce` -> `resource`
> * [x]  Check `cartesian` -> `Cartesian'
> * [x]  Check `procesing` -> `processing`
> * [x]  Check `accomodate` -> `accommodate`
> * [x]  Can you check for consistency with either American or European (Brittish/Irish) English? E.g. You use `optimized`, `vectorized`, and `specialized` which favors American English, whereas the use of Aluminium over Aluminum favours Brittish English. Either is fine but it is best to be consistent.
> * [x]  Check `Overview of experiment setup`, this should perhaps be improved to something like `Overview of the experimental setup`
> * [x]  Check `Seen from side` -> `Seen from the side`
> * [x]  Check `ultrasonic raw data` -> `raw ultrasonic data`
