### 2022-08-08 #4185 @Kevin-Mattheus-Moerman
@mh-skjelvareid Now that all reviewers are happy with both the software and the paper we can proceed to process this for acceptance in JOSS. Here are some action items for you:

- [ ] Please archive a copy of the software on ZENODO. You can do this manually if you like but some find [these automated steps](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) useful.
- [ ] Please check and edit the ZENODO archive so that the title and the authors match those of the paper. Also check that the license is correct. Furthermore we recommend that you add your ORCID information to the ZENODO archive too.
- [ ] Once you have the archived version created, please report back here with the archive DOI.
- [ ] The current version tag is `v1.1.0`, is this still accurate or should we update this? Note the automated approach above may cause you to issue an updated version so please inform us what the version tag should be.

Thanks.
