### 2022-08-08 dacb767
Update .gitignore and add files regarding radar processing
