+++ experimental/psm_spec_input/c_psm_2d_batch.m
@@ -1,0 +1,155 @@
+%   psm_2d_batch - batch migration using PSM and preprocessed data
+%
+%   Note that this script uses two local functions,
+%   loadAndConcatenateBlocks() and calc_spec_kx(), defined at the end of
+%   the file.
+%
+%   2021-09-16  MHS - Initial version
+%
+
+%% Reset
+close all
+clearvars
+
+%% Define data paths, get all file names
+% Files are assumed to be named so that they are sorted in the same order
+% that they were originally generated (i.e. the first file in a
+% alphabetically sorted list contains the first part of the scan, etc.).
+
+% Update this path if re-processing on your local computer
+dataPath = 'E:\Synaptus\RadarAntarcticaData\Preprocessed\';
+
+%% Create folder for saving output (if not already in place)
+imagePath = [dataPath 'Images\'];
+if not(exist(imagePath,'dir'))
+    mkdir(imagePath)
+end
+
+%% Constants
+c_medium_orig = 168913914;          % Original wave speed (from "settings")
+c_medium = c_medium_orig*1.45;      % Modified wave speed, used in processing, seems better(?)
+tDelay = 0;
+tMax = 4.6016e-05;
+thick = (c_medium/2)*tMax;
+
+%% Get file names
+matFiles = dir([dataPath '*.mat']);
+matFiles = {matFiles.name};
+nFiles = length(matFiles);
+
+%% Process data - concatenate neigboring blocks, use image from middle block
+
+for ii = 1:nFiles
+    % Set file indices for neighboring blocks, handle edge cases
+    fileInd = [-1 0 1] + ii;
+    fileInd = fileInd((fileInd>=1) & (fileInd<=nFiles));
+    
+    % Concatenate blocks and Fourier transform along x axis
+    [Pox,index,omega,xStep] = loadAndConcatenateBlocks(dataPath, matFiles(fileInd));
+    [Pokx,kx] = calc_spec_kx(Pox,xStep);
+    
+    % Create focused image (migration)
+    disp(['Creating focused image based on files numbered ' num2str(fileInd)])
+    [im,xIm,zIm] = psm_spec_input(Pokx,omega,kx,xStep,tDelay,c_medium,thick);
+    
+    % Crop image to second block (except for left edge)
+    for ll = 1:length(im)
+        if ii == 1
+            im{ll} = im{ll}(:,index(1,1):index(1,2));   % Only two blocks, keep left
+        else
+            im{ll} = im{ll}(:,index(2,1):index(2,2));   % Crop to second block
+        end
+    end
+
+    % Save focused image
+    [~,fileBase,~] = fileparts(matFiles{ii});
+    imageFileName = [imagePath fileBase '_focused.mat'];
+    save(imageFileName,'im','xIm','zIm','c_medium')
+
+end
+
+function [Pox_concat,index,omega,xStep] = loadAndConcatenateBlocks(matFileFolder,matFiles)
+% loadAndConcatenateBlocks - load preprocessed files, tile horizontally
+%
+% Usage:
+% [Pox_concat,index,omega,xStep] = loadAndConcatenateBlocks(matFileFolder,matFiles)
+%
+% Input parameters:
+% matFileFolder -   path to folder containing matfiles
+% matFiles      -   cell array with names to matfiles containing 'Pox',
+%                   'omega', 'xStep' variables
+%
+% Output parameters:
+% Pox_concat    - blocks concatenated horizontally
+% index         - indices to start and end of each block (horizontally),
+%                 size [nFiles,2]
+% omega         - vector of omega values, corresponding to vertical axis of
+%                 Pox_concat
[…]
