+++ experimental/array_psm_planewave/array_psm_planewave.m
@@ -1,202 +1,0 @@
-function varargout = array_psm_planewave(ptxa,fs,txDelay,cc,thick,fLow,fHigh,aPitch,varargin)
-% array_psm_planewave - Phase Shift Migration for linear arrays with steered plane waves
-%
-%   NOTE: This function only handles the special case of steered plane
-%   waves, while the array_psm() function handles arbitrary transmitted
-%   wave shapes, and also returns a wider field of view. The array_psm
-%   function is therefore recommended over this one. 
-%
-%   Usage:
-%   [im,...] = array_psm_planewave(ptxa,fs,txDelay,cc,thick,fLow,fHigh,aPitch,...)
-%
-%   Input parameters:
-%   ptxa    -    ultrasonic data, dimensions (t,x,angle). It is assumed
-%                that there is no "measurement delay", i.e. the data starts
-%                at t=0, when the first pulse is fired. 
-%   fs      -    Sampling frequency [Hz]
-%   txDelay -    2D array of pulse time delays [s], dimensions (x, angle)
-%   cc      -    Vector of compressional wave velocities for each layer [m/s]
-%   thick   -    Vector of thickness for each layer [m/s]
-%                  (set last boundary equal to end of ROI)
-%   fLow    -    Lower limit of transducer frequency band [Hz]
-%   fHigh   -    Higher limit of transducer frequency band [Hz]
-%   aPitch  -    array pitch (distance between array element centers) [m]
-%
-%   Optional parameter-value input pairs:
-%   tFftMult    -    multiplier for t-axis FFT size. Default: 1
-%   xFftMult    -    multiplier for x-axis FFT size. Default: 2
-%   zResMult    -    multiplier for final image z resolution. Default: 2
-%   wf          -    2-way waveform, for matched filt. Default: 1 (no filt.)
-%   skipLayers  -    vector w/indices of layers to skip in processing 
-%                   (empty output). Default: [] (all layers included).
-%
-%   Output parameters (varargout):
-%   im          -    Reconstructed image as cell array - one cell per layer
-%   xIm         -    x positions for pixels in im
-%   zIm         -    z positions for pixels in im
-%                       (cell array, different z res. for each layer)
-%
-%   Note: The default z resolution of the output image of layer L is
-%       dz = (1/zResMult)*(cc(L)/2)/(fHigh-fLow).
-%   This is done to minimize the number of z depths processed for each
-%   layer and speed up the imaging. If this resolution is too coarse, 
-%   try resampling the image to a higher resolution (see resample, interp)
-%   or increase the zResMult parameter.  
-%
-%   The derivation of the PSM algorithm is described in sections 3.5 and 5.1
-%   the PhD thesis "Synthetic aperture ultrasound imaging with application to
-%   interior pipe inspection" (2012) by Martin H. Skjelvareid, University of
-%   Tromsø, Norway. However, ahe adaptation to array imaging is not 
-%   included in the PhD thesis.
-%
-%   LICENSE DISCLAIMER:
-%   This program is free software: you can redistribute it and/or modify
-%   it under the terms of the GNU General Public License as published by
-%   the Free Software Foundation, either version 3 of the License, or
-%   (at your option) any later version.
-%
-%   This program is distributed in the hope that it will be useful,
-%   but WITHOUT ANY WARRANTY; without even the implied warranty of
[…]
