+++ experimental/psm_spec_input/psm_spec_input.m
@@ -1,134 +1,0 @@
-function varargout = psm_spec_input(Pokx,omega,kx,xStep,tDelay,cc,thick,varargin)
-% psm_spec_input - Phase Shift Migration with spectral input
-%
-%   Usage:
-%   [im,...] = psm_spec_input(Pokx,omega,kx,tDelay,cc,thick,...)
-%
-%   Input parameters:
-%   Pokx     -  MxN matrix of spectral data, dimensions (omega,kx).
-%   omega   -   vector, length M, with omega coordinates.
-%               Omega values are assumed to be positive, corresponding to 
-%               the passband of the transducer / antenna / signal. 
-%   kx      -   vector, length N, with kx coordinates.
-%               Kx values are usually both positive and negative (symmetry
-%               around zero), but this is not a requirement.
-%   tDelay  -    Time delay between pulse transmission and measurement [s]
-%   cc      -    Vector of compressional wave velocities for each layer [m/s]
-%   thick   -    Vector of thickness for each layer [m/s]
-%                  (set last boundary equal to end of ROI)
-%
-%   Output parameters (varargout):
-%   im          -    Reconstructed image as cell array - one cell per layer
-%   xIm         -    x positions for pixels in im
-%   zIm         -    z positions for pixels in im
-%                       (cell array, different z res. for each layer)
-%
-%   This is a version of the PSM algorithm assuming that the data has
-%   already been Fourier transformed along the time dimension, and that the
-%   "passband" of the transducer/antenna/signal has been extracted. The
-%   motivation for doing this is that extraction of the passband represents
-%   a reduction in data size, making it easier to fit a large dataset into
-%   memory.
-%
-%   The derivation of the PSM algorithm is described in sections 3.5 and 5.1
-%   the PhD thesis "Synthetic aperture ultrasound imaging with application to
-%   interior pipe inspection" (2012) by Martin H. Skjelvareid, University of
-%   Tromsø, Norway.
-%
-%   LICENSE DISCLAIMER:
-%   This program is free software: you can redistribute it and/or modify
-%   it under the terms of the GNU General Public License as published by
-%   the Free Software Foundation, either version 3 of the License, or
-%   (at your option) any later version.
-%
-%   This program is distributed in the hope that it will be useful,
-%   but WITHOUT ANY WARRANTY; without even the implied warranty of
-%   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-%   GNU General Public License for more details.
-%
-%   You should have received a copy of the GNU General Public License
-%   along with this program.  If not, see <http://www.gnu.org/licenses/>.
-%
-%   2021-09-16  MHS - Initial version
-%
-%   Copyright (C) 2021  Martin H. Skjelvareid
-%   martin.hansen.skjelvareid@gmail.com
-
-
-%% Dimension testing 
-[nOmega,nKx] = size(Pokx);
-if length(omega) ~= nOmega
-    error('Length of omega vector must correspond to number of rows in Pokx')
-end
-if length(kx) ~= nKx
-    error('Length of kx vector must correspond to number of columns in Pokx')
-end
-
-%% Calculate dependent variables
-nL = length(thick);                                 % Number of layers
-zIF = cumsum([0; thick(:)]);                        % z position of interfaces
-
[…]
