### 2022-08-08 #4185 @mh-skjelvareid
@Kevin-Mattheus-Moerman , see replies below:

> @mh-skjelvareid Now that all reviewers are happy with both the software and the paper we can proceed to process this for acceptance in JOSS. Here are some action items for you:
>
> * [x]  Please archive a copy of the software on ZENODO. You can do this manually if you like but some find [these automated steps](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) useful.
> * [x]  Please check and edit the ZENODO archive so that the title and the authors match those of the paper. Also check that the license is correct. Furthermore we recommend that you add your ORCID information to the ZENODO archive too.
> * [x]  Once you have the archived version created, please report back here with the archive DOI.

All done. The DOI is [10.5281/zenodo.6974194](https://doi.org/10.5281/zenodo.6974194)
I've added the DOI badge at the top of the ´README.md´ file: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6974195.svg)](https://doi.org/10.5281/zenodo.6974195)

I think I got all the details right, but please check that I did. My chosen license is GNU GPL v3, but this wasn't available in the drop-down menu of possible licenses, so I chose the closest one, "GNU General Public License v2.0 or later"

> * [x]  The current version tag is `v1.1.0`, is this still accurate or should we update this? Note the automated approach above may cause you to issue an updated version so please inform us what the version tag should be.

I believe that until now, the version has been v1.0 ("Initial release"), but I've created another release in order to get the Zenodo DOI, so the current (latest) version is v1.1. Does it look OK?
