### 2022-08-08 #4185 @Kevin-Mattheus-Moerman
@mh-skjelvareid Odd that the GPL v3 license does not show up, I just checked and saw the same. Perhaps instead of picking a license which is the "closest one" I recommend you select the option "Other (open)".
