### 2022-08-08 #4185 @mh-skjelvareid
Thanks a lot, @Kevin-Mattheus-Moerman, for stepping in yourself and making the publication happen (and for giving many useful tips along the way) - it really means a lot! :) And thanks again also to @emanuelhuber for generously using your time to review the software.

I've found the whole process very interesting, and (with the exception of problems finding reviewers) very smooth and more "human" than similar processes I've been through before. Both the idea and the implementation of JOSS are really good! I will be signing up to be a reviewer myself.
