### 2022-08-09 #4185 @Kevin-Mattheus-Moerman
@mh-skjelvareid you are welcome @mh-skjelvareid. My apologies that the review to much longer than is normal for JOSS. Since COVID finding reviewers has been a challenges (as most people were/are behind on lots of work) as well as getting them to commit to a speedy review process.

Great that you want to help review for JOSS.

Congrats again!
